+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# HEAD
+# v0.4.5 (2019-05-12)
 
 - feat(task editor): Set default value when creating interval tasks
 - feat(task editor): Press Escape to close editor